Files are added. Scripts are fixed
