+++ Homework2/requirements.txt
@@ -1,0 +1,7 @@
+click==8.0.1
+colorama==0.4.4
+Flask==2.0.1
+itsdangerous==2.0.1
+Jinja2==3.0.1
+MarkupSafe==2.0.1
+Werkzeug==2.0.1
